bug#179831 clean compiler warning     [bug number ]179831     [root cause] clean compiler warning     [changes] clean compiler warning     [side effects] no     [reviewers] steve zhan
	modified:   arch/arm/mach-sc/pin_switch.c

+++ arch/arm/mach-sc/pin_switch.c
@@ -58,7 +58,7 @@ static int pin_switch_debug_set(void *data, u64 val)
 static int pin_switch_debug_get(void *data, u64 * val)
 {
 	struct sci_pin_switch *p = data;
-	*val = read_write_pin_switch(1, val, p);
+	*val = read_write_pin_switch(1, (int)val, p);
 	return 0;
 }
